requirements.txt
anthropic>=0.42.0
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
markupsafe>=2.1.0
